toki: fix drawing on real hardware
toki will only update the screen when you write to 0x0a0040. If
you do this outside of vblank it will trigger some snow effects on
screen. To avoid this as much as possible we are doing

- when in PSUB mode, force an update after each print_* dsub
- when in RSUB mode, enable ints, do the update in the vblank int

mame has no logic for 0x0a0040 and just seems to force the screen
update every vblank.