Rework memory_data_test
This changes the logic so it writes out the entire test region,
then goes back and verifies it.

This fixes a false failure I was seeing on the palette ram of
wwfsstar on real hardware.  I'm assuming this was being caused
by the test doing really fast write then re-read, but at the
same time the rendering logic of the board also needs to access
the palette ram.