Implement memory_tests_handler and hookup wwfsstar to it
You pass memory_tests_handler data with details about the memory
(start address, size, etc) and it will deal with running all the
different memory tests against it and setting the proper error
code when returning.  This gets rid of a bunch of dupe code that
was needed from each memory type having to setup and call each of
the different ram tests.