gaiden: don't do march ram tests on bg/txt ram
bg/txt ram both pass multiple test loops of tests but will
eventually fail the march test.  This has been confirmed on
multiple boards.  I'm assuming this is the result of the cpu and
the custom chips both accessing the ram at the same time and
causing a conflict.  The way the march test works must be tickling
this.