rom-inject-crc-mirror: add reverse (-r) cli option
On some CPUs the reset vector is at the end of address space.  In
this case the rom will also be mapped into the space.  Because
of this the last copy of the diag in the rom file will be the
active one.  The reverse option will cause crc32/mirror data to
be written out backwards through the diag mirrors in the rom file.

ie: The last copy will get the crc32 data and be mirror 0.