Change mad version number to be based on last git commit date
gen-version-asm-file script is used to generate a version.asm based
on a version.asm.in template, where the date (YYYY-MM-DD) of the last
git commit is used for the version number of mad.  If there are
uncommitted changes then the version is appended with "D" to indicate
its a dirty build (ie: 2025-10-20D)

version.asm is excluded from git and is regenerated every build.