Add test konami cpu support and crimfght machine
I've been helping get the konami cpu (6809 based) cpu support
added to vasm.  This commit mainly contains code related to
testing that cpu is assembling ok and works on hardware. The
current focus has been around figuring out whats going on with
TFR/EXG instructions on real hardware.

If you are wanting to mess with this you will need to use a daily
build of vasm.