rom-inject-crc-mirror: add -M option
This adds a copy of the mirror number at the first byte of each
mirror and is only valid with the -r option.  This is useful when
testing prog rom banking, where the prog bank size is only 0x2000
while the diag rom mirrors are 0x4000 in size.  You have access to
the diag mirror number when either half of the mirror is banked in.