6309+machines: rework memory output test
The previous code was slightly bugged in that it should have been
doing `lda [,x]` which would cause the contents of memory location
$ffff to be loaded into `a` if the memory location pointed at by
`x` had dead output.

While having a test rig setup I tested the 6809 memory output code
on a 6309 machine (contra) and it worked fine.  Instead of fixing
the lda I decided to switch to the 6309 code so both CPUs where using
the same code.

This also means being able to remove the reset section/shenanigans
from 6309 machines.