konami2: add MT_FLAG_INTERLEAVED option to memory test handler
Palette ram on konami2 based boards have 2 ram chips, where one chip
is for even addresses and the other one is odd addresses.

When this flag is enabled the memory test handler will also test an
odd address for output, write and address tests to ensure we are
properly testing the 2nd chip.