z80: use WORK_RAM_CHIP* defines for work ram tests
Work ram will not always have a dedicated ram chip and instead
can share it with other parts of the board's logic (ie tile/sprite
data).  When testing work ram we need to test the entire ram chip
and not just the part of the chip work ram is in.  The work ram
tests will now use WORK_RAM_CHIP* defines, provided by the machine,
for the location/size/address lines to test.