z80: add values edit handler; mitchell: add sprite viewer
Values edit handler allows passing in an array of variables, masks
and some other info, then allows the user to modify the values.
Whenever one of the values is changed it will trigger a callback
function to allow doing stuff with the variables/values.

A sprite viewer for mitchell boards was created using this handler.