konami2: print the error'd address on failed work ram tests
Printing the actual/expected values clobblers the error'd address
in the X register.  However there is a quark with the konami2 CPU
where the DP register is actually 16bits.  So we can `tfr x, dp`,
print actual/expected values, then restore and print the error'd
address.  Note: MAME doesn't support tfr to the DP register, so it
will always print the error'd work ram address as 0x0000.

Additionally the konami2 CPU doesn't support tfr to the D register,
so we have to jump through some hoops to get the error'd address in
there for printing.